# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Giang Minh Viet Apis",
    "version": "1.0.0",
    "title": "Giang Minh Viet Apis",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:4000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/inquiry/": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "item per Page",
            "in": "query",
            "description": "Max Inquiry can be display in one page",
            "required": false
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Page number",
            "required": false
          },
          {
            "name": "orderType",
            "in": "query",
            "description": "Order type to display: 'newest' or 'oldest' or null",
            "required": false
          }
        ],
        "tags": [
          "Inquiry Api"
        ],
        "summary": "Get Inquiry",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/inquiry"
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "require": true,
            "schema": {
              "type": "object",
              "properties": {
                "customer_name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "product_name": {
                  "type": "string"
                },
                "product_id": {
                  "type": "integer"
                },
                "product_link": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "tags": [
          "Inquiry Api"
        ],
        "summary": "Create Inquiry",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "insertedId": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/inquiry/{id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          }
        ],
        "tags": [
          "Inquiry Api"
        ],
        "summary": "Get inquiry by Id",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/inquiry"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          }
        ],
        "tags": [
          "Inquiry Api"
        ],
        "summary": "Delete Inquiry by ID",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/success"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/inquiry/get-by-customer-name/{name}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "type": "string",
            "in": "path",
            "required": true
          }
        ],
        "tags": [
          "Inquiry Api"
        ],
        "summary": "Get Inquiry by Name",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/inquiry"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "inquiry": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "customer_name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "product_id": {
          "type": "integer"
        },
        "quantity": {
          "type": "integer"
        }
      }
    },
    "error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "success": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}